Create from def "enable_source_repos" repodict function
It is one of the solution how to move a code originated from plugins-core.
